Move NCCLDevCommManager registry into its own source
The cross-DSO fix needs an out-of-line get() definition in libtorch_cuda,
but placing it in nccl_devcomm_bridge.cpp couples registry ownership to
registration-hook installation.
Move the definition into nccl_devcomm_manager.cpp,
retain the bridge for hooks only, and register the new source in the shared build list.

Authored with GPT-5.6.
