+++ build_variables.bzl
@@ -811,6 +811,7 @@ libtorch_cuda_distributed_extra_sources = [
     "torch/csrc/distributed/c10d/symm_mem/NCCLSymmetricMemory.cu",
     "torch/csrc/distributed/c10d/symm_mem/nccl_extension.cu",
     "torch/csrc/distributed/c10d/symm_mem/nccl_devcomm_bridge.cpp",
+    "torch/csrc/distributed/c10d/symm_mem/nccl_devcomm_manager.cpp",
     "torch/csrc/distributed/c10d/symm_mem/ops/nccl_reduce_scatter_offset.cu",
     "torch/csrc/distributed/c10d/symm_mem/ops/nccl_all_to_all_nd.cu",
     "torch/csrc/distributed/c10d/symm_mem/intra_node_comm.cpp",
